fix: comprehensive crate fixes and eBPF improvements
- Fixed IPv4/IPv6 key mismatch in XDP BPF code (u32 → u128)
- Fixed WAL discover_max_lsn to return Option<u64> for empty state
- Fixed WAL replay EOF detection with peek-based approach
- Fixed IPv6 subnet_key_v6 to zero host bits for /64
- Added #[allow(unsafe_code)] with safety comments for aya Pod impls
- Removed unused imports and variables across crates
- Fixed BlockReason import conflicts in storage crate
- All tests passing, zero warnings

+++ src/enforcement/mod.rs
@@ -12,7 +12,7 @@ use crate::storage::{Store, Value, BlockState, BlockReason, IpRecord};
 use crate::metrics::Metrics;
 use crate::util::BoundedVecDeque;
 use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::net::IpAddr;
 use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
@@ -223,7 +223,7 @@ impl EnforcementService {
 
                 rec.block_state = BlockState::Blocked { reason: block_reason, since_ns: now_ns };
 
-                self.store.insert(key, Value::IpRecord(rec), ttl_secs, self.store.ram_bytes() as usize)
+                self.store.insert(key, Value::IpRecord(rec), ttl_secs, self.store.ram_bytes())
                     .map_err(|e| EnforcementError::Storage(e.to_string()))
             }
             EnforceAction::Unblock => {
@@ -236,7 +236,7 @@ impl EnforcementService {
                 
                 if let Some(Value::IpRecord(mut rec)) = self.store.get(&cmd.ip) {
                     rec.block_state = BlockState::Clean;
-                    self.store.insert(cmd.ip, Value::IpRecord(rec), None, self.store.ram_bytes() as usize)
+                    self.store.insert(cmd.ip, Value::IpRecord(rec), None, self.store.ram_bytes())
                         .map_err(|e| EnforcementError::Storage(e.to_string()))
                 } else {
                     Ok(()) // Not tracked = already unblocked
